Refactor GraphQL query validation  (#2073)
* Refactor GraphQLDocumentParser

- Extract exceptions
- Extract some utils extensions
- Extract pos parse validations

* Add tests for variables validation

* Add fragment validation tests

* More tests

* Fix typo

(cherry picked from commit 23bd3b597cefdb3223df93cecbfc2e74a9291f18)
