+++ apollo-compiler/src/test/validation/graphql/FragmentDuplication.error
@@ -1,0 +1,1 @@
+/FragmentDuplication.graphql: There can be only one fragment named `HeroFragment`
