+++ apollo-compiler/src/test/validation/graphql/FragmentUnknownType.graphql
@@ -1,0 +1,9 @@
+query TestQuery {
+  hero {
+    name
+  }
+}
+
+fragment HeroFragment on Banana {
+  name
+}
