+++ apollo-compiler/src/test/validation/graphql/InlineFragmentCanNotBeSpread.error
@@ -1,0 +1,7 @@
+Failed to parse GraphQL file /InlineFragmentCanNotBeSpread.graphql (3:11)
+Fragment cannot be spread here as objects of type `Character` can never be of type `Mutation`
+----------------------------------------------------
+[2]:  hero {
+[3]:    ... on Mutation {
+[4]:
+----------------------------------------------------
