+++ apollo-compiler/src/test/validation/graphql/InlineFragmentCanNotBeSpread.graphql
@@ -1,0 +1,7 @@
+query TestQuery {
+  hero {
+    ... on Mutation {
+
+    }
+  }
+}
