+++ apollo-compiler/src/test/validation/graphql/InlineFragmentUnknonwType.graphql
@@ -1,0 +1,6 @@
+query TestQuery {
+  hero {
+    ... on Banana {
+    }
+  }
+}
