+++ apollo-compiler/src/test/validation/graphql/QueryUnknowField.graphql
@@ -1,0 +1,5 @@
+query TestQuery {
+  hero12 {
+    name
+  }
+}
