+++ apollo-compiler/src/test/validation/graphql/VariableIsOptional.graphql
@@ -1,0 +1,5 @@
+query TestQuery($episode: Episode) {
+  hero(episode: $episode) {
+    name
+  }
+}
