+++ apollo-compiler/src/test/validation/graphql/VariableMissing.graphql
@@ -1,0 +1,5 @@
+query TestQuery {
+  hero(episode: $episode) {
+    name
+  }
+}
