+++ apollo-compiler/src/test/validation/graphql/VariableTypeMismatch.error
@@ -1,0 +1,7 @@
+Failed to parse GraphQL file /VariableTypeMismatch.graphql (2:16)
+Variable `episode` of type `Int` used in position expecting type `Episode!`
+----------------------------------------------------
+[1]:query TestQuery($episode: Int) {
+[2]:  hero(episode: $episode) {
+[3]:    name
+----------------------------------------------------
