+++ apollo-compiler/src/test/validation/graphql/VariableTypeMismatch.graphql
@@ -1,0 +1,5 @@
+query TestQuery($episode: Int) {
+  hero(episode: $episode) {
+    name
+  }
+}
